fix issue with fd

+++ srcs/utils/ft_free_2.c
@@ -3,14 +3,30 @@
 /*                                                        :::      ::::::::   */
 /*   ft_free_2.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: juliensarda <juliensarda@student.42.fr>    +#+  +:+       +#+        */
+/*   By: jsarda <jsarda@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/17 12:02:40 by jtaravel          #+#    #+#             */
-/*   Updated: 2024/07/19 22:08:50 by juliensarda      ###   ########.fr       */
+/*   Updated: 2024/07/22 15:32:00 by jsarda           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
+
+void	ft_close_fd(t_data *data)
+{
+	while (data)
+	{
+		if (data->fdin != -1 && data->fdin != 0)
+			close(data->fdin);
+		if (data->fdout != -1 && data->fdout != 1)
+		{
+			if (data->next && data->next->fdout != data->fdout)
+				return ;
+			close(data->fdout);
+		}
+		data = data->next;
+	}
+}
 
 void	free_hd_file(t_data **data, int mode)
 {
@@ -25,7 +41,7 @@ void	free_hd_file(t_data **data, int mode)
 				free(tmp_data->tmpfile_hd);
 			tmp_data->tmpfile_hd = NULL;
 		}
-		else if (mode == 2 && tmp_data->is_hd)
+		else if (mode == 2 && tmp_data->is_hd && tmp_data->tmpfile_hd)
 			unlink(tmp_data->tmpfile_hd);
 		tmp_data = tmp_data->next;
 	}
@@ -38,6 +54,7 @@ void	free_child(t_data *data, t_shell *shell, int exit_status)
 	free(data->path);
 	data->path = NULL;
 	free_hd_file(&data, 1);
+	ft_close_fd(data);
 	ft_clear_datas(&(shell->datas));
 	free(shell);
 	exit(exit_status);
